SDI-2223-505-sesión10-Ejercicio-complementario1.

+++ musicstoreapp/routes/api/songsAPIv1.0.js
@@ -1,4 +1,5 @@
 const {ObjectId} = require("mongodb");
+const songsRepository = require("../../repositories/songsRepository");
 module.exports = function (app, songsRepository, usersRepository) {
     app.get("/api/v1.0/songs", function (req, res) {
         let filter = {};
@@ -39,18 +40,18 @@ module.exports = function (app, songsRepository, usersRepository) {
         try {
             let songId = ObjectId(req.params.id)
             let filter = {_id: songId}
-            songsRepository.deleteSong(filter, {}).then(result => {
-                if (result === null || result.deletedCount === 0) {
-                    res.status(404);
-                    res.json({error: "ID inválido o no existe, no se ha borrado el registro."});
-                } else {
-                    res.status(200);
-                    res.send(JSON.stringify(result));
-                }
-            }).catch(error => {
-                res.status(500);
-                res.json({error: "Se ha producido un error al eliminar la canción."})
-            });
+            let isOwner = isSongOwner(res.user, songId);
+            if (isOwner) {
+                songsRepository.deleteSong(filter, {}).then(result => {
+                    if (result === null || result.deletedCount === 0) {
+                        res.status(404);
+                        res.json({error: "ID inválido o no existe, no se ha borrado el registro."});
+                    }
+                });
+            } else {
+                res.status(400);
+                res.json({error: "Debes ser el dueño de la canción para poder eliminarla."});
+            }
         } catch (e) {
             res.status(500);
             res.json({error: "Se ha producido un error, revise que el ID sea válido."})
@@ -61,18 +62,30 @@ module.exports = function (app, songsRepository, usersRepository) {
         try {
             let song = {title: req.body.title, kind: req.body.kind, price: req.body.price, author: req.session.user}
             // Validar aquí: título, género, precio y autor.
-            songsRepository.insertSong(song, function (songId) {
-                if (songId === null) {
-                    res.status(409);
-                    res.json({error: "No se ha podido crear la canción. El recurso ya existe."});
-                } else {
-                    res.status(201);
-                    res.json({message: "Canción añadida correctamente.", _id: songId})
-                }
-            });
+            let error_msg = "";
+            if (typeof song.title === "undefined" || song.title === null || song.title.length === 0)
+                error_msg = "Título incorrecto, la longitud ha de ser mayor a 0";
+            if (typeof song.kind === "undefined" || song.kind === null)
+                error_msg = "Tipo incorrecto";
+            if (typeof song.price === "undefined" || song.price === null || parseFloat(song.price) < 0)
+                error_msg = "Precio incorrecto, ha de ser positivo";
+            if (error_msg.length !== 0) {
+                res.status(403);
+                res.json({error: "Se ha introducido mal algún dato. Error: " + error_msg}); // Error detectado
+            } else {
+                songsRepository.insertSong(song, function (songId) {
+                    if (songId === null) {
+                        res.status(409);
+                        res.json({error: "No se ha podido crear la canción. El recurso ya existe."});
+                    } else {
+                        res.status(201);
+                        res.json({message: "Canción añadida correctamente.", _id: songId});
+                    }
+                });
+            }
         } catch (e) {
             res.status(500);
-            res.json({error: "Se ha producido un error al intentar crear la canción: " + e})
+            res.json({error: "Se ha producido un error al intentar crear la canción: " + e});
         }
     });
 
@@ -85,27 +98,42 @@ module.exports = function (app, songsRepository, usersRepository) {
             let song = {author: req.session.user}
             if (typeof req.body.title !== "undefined" && req.body.title !== null)
                 song.title = req.body.title;
+            else {
+                res.status(400);
+                return res.json({error: "Datos incorrectos"});
+            }
             if (typeof req.body.kind !== "undefined" && req.body.kind !== null)
                 song.kind = req.body.kind;
+            else {
+                res.status(400);
+                return res.json({error: "Datos incorrectos"});
+            }
             if (typeof req.body.price !== "undefined" && req.body.price !== null)
                 song.price = req.body.price;
-            songsRepository.updateSong(song, filter, options).then(result => {
-                if (result === null) {
-                    res.status(404);
-                    res.json({error: "ID inválido o no existe, no se ha actualizado la canción."});
-                }
-                //La _id No existe o los datos enviados no difieren de los ya almacenados.
-                else if (result.modifiedCount == 0) {
-                    res.status(409);
-                    res.json({error: "No se ha modificado ninguna canción."});
-                } else {
-                    res.status(200);
-                    res.json({message: "Canción modificada correctamente.", result: result})
-                }
-            }).catch(error => {
-                res.status(500);
-                res.json({error: "Se ha producido un error al modificar la canción."})
-            });
+            else {
+                res.status(400);
+                return res.json({error: "Datos incorrectos"});
+            }
+            let isOwner = isSongOwner(res.user, songId);
+            if (isOwner) {
+                songsRepository.updateSong(song, filter, options).then(result => {
+                    if (result === null) {
+                        res.status(404);
+                        res.json({error: "ID inválido o no existe, no se ha actualizado la canción."});
+                    }
+                    //La _id No existe o los datos enviados no difieren de los ya almacenados.
+                    else if (result.modifiedCount === 0) {
+                        res.status(409);
+                        res.json({error: "No se ha modificado ninguna canción."});
+                    } else {
+                        res.status(200);
+                        res.json({message: "Canción modificada correctamente.", result: result})
+                    }
+                });
+            } else {
+                res.status(400);
+                res.json({error: "No es el dueño de la canción."});
+            }
         } catch (e) {
             res.status(500);
             res.json({error: "Se ha producido un error al intentar modificar la canción: " + e})
@@ -150,4 +178,13 @@ module.exports = function (app, songsRepository, usersRepository) {
             });
         }
     });
+
+    function isSongOwner(user, songId) {
+        let filter = {_id: songId};
+        return songsRepository.findSong(filter, {}).then(song => {
+            if (song.author === user)
+                return true;
+            return false;
+        });
+    }
 }
